feat: stream-b08-provider-ownership-cleanup-003 - Own remaining shared failure parsers in provider packages
