+++ pkg/workers/provider/codex/failure.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	provider "github.com/portpowered/infinite-you/pkg/workers/provider"
+	"github.com/portpowered/infinite-you/pkg/workers/provider/codex/exitfailure"
 )
 
 const unknownTerminalFailureMessage = "Codex reported a terminal error."
@@ -60,7 +61,7 @@ func shouldSelectTerminalFailure(selectedRecognized, candidateRecognized bool) b
 }
 
 func classifyTerminalMessage(nativeType, message, threadID string) (TerminalFailure, bool) {
-	parsed := provider.ParseCodexProviderFailure(provider.CommandResult{
+	parsed := exitfailure.ParseExitFailure(exitfailure.ExitFailureInput{
 		ExitCode: 1,
 		Stderr:   []byte("ERROR: " + strings.TrimSpace(message)),
 	})
@@ -68,7 +69,9 @@ func classifyTerminalMessage(nativeType, message, threadID string) (TerminalFail
 	if !recognized {
 		parsed.Message = unknownTerminalFailureMessage
 	}
-	providerErr := provider.NewProviderErrorFromResult(parsed, nil)
+	providerErr := provider.NewProviderErrorFromResult(provider.ProviderFailureResult{
+		Reason: parsed.Reason, Message: parsed.Message,
+	}, nil)
 	decision := provider.WorkFailureDecisionFromProviderError(providerErr)
 	return TerminalFailure{
 		Type: parsed.Reason, Message: parsed.Message, Retryable: decision.Retryable,
