+++ pkg/workers/provider/helpers.go
@@ -9,7 +9,6 @@ import (
 	"sort"
 	"strings"
 	"time"
-	"unicode"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	workerprocess "github.com/portpowered/infinite-you/pkg/workers/process"
@@ -579,338 +578,3 @@ func workerEventExitCode(exitCode int, present bool, includeZero bool) *int {
 	return &exitCodeCopy
 }
 
-func lastSafeOpenCodeUnknownExcerpt(streams []string) (string, bool) {
-	var last string
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			trimmed := strings.TrimSpace(line)
-			if failure, ok := decodeOpenCodeStructuredFailure(trimmed); ok {
-				if excerpt, safe := safeOpenCodeFailureDetail(failure.Message); safe {
-					last = excerpt
-				}
-				continue
-			}
-			if !openCodeErrorTextLine(trimmed) {
-				continue
-			}
-			if excerpt, safe := safeOpenCodeFailureDetail(trimmed); safe {
-				last = excerpt
-			}
-		}
-	}
-	return last, last != ""
-}
-
-func lastOpenCodeStructuredFailure(streams []string) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
-	var found bool
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			failure, ok := decodeOpenCodeStructuredFailure(strings.TrimSpace(line))
-			if !ok {
-				continue
-			}
-			reason, recognized := classifyOpenCodeStructuredFailure(failure)
-			if !recognized {
-				continue
-			}
-			last = ProviderFailureResult{
-				Reason:  reason,
-				Message: openCodeFailureMessage(reason, failure.Message),
-			}
-			found = true
-		}
-	}
-	return last, found
-}
-
-func decodeOpenCodeStructuredFailure(line string) (opencodeStructuredFailure, bool) {
-	if !strings.HasPrefix(line, "{") {
-		return opencodeStructuredFailure{}, false
-	}
-	var envelope struct {
-		Type       string `json:"type"`
-		Name       string `json:"name"`
-		Code       string `json:"code"`
-		Status     int    `json:"status"`
-		StatusCode int    `json:"statusCode"`
-		Message    string `json:"message"`
-		Error      *struct {
-			Type       string `json:"type"`
-			Name       string `json:"name"`
-			Code       string `json:"code"`
-			Status     int    `json:"status"`
-			StatusCode int    `json:"statusCode"`
-			Message    string `json:"message"`
-			Data       *struct {
-				Code       string `json:"code"`
-				Status     int    `json:"status"`
-				StatusCode int    `json:"statusCode"`
-				Message    string `json:"message"`
-			} `json:"data"`
-		} `json:"error"`
-	}
-	if err := json.Unmarshal([]byte(line), &envelope); err != nil {
-		return opencodeStructuredFailure{}, false
-	}
-
-	failure := opencodeStructuredFailure{
-		Name:       envelope.Name,
-		Type:       envelope.Type,
-		Code:       envelope.Code,
-		StatusCode: firstNonZero(envelope.StatusCode, envelope.Status),
-		Message:    envelope.Message,
-	}
-	if envelope.Error != nil {
-		failure.Name = firstNonEmpty(envelope.Error.Name, failure.Name)
-		failure.Type = firstNonEmpty(envelope.Error.Type, failure.Type)
-		failure.Code = firstNonEmpty(envelope.Error.Code, failure.Code)
-		failure.StatusCode = firstNonZero(envelope.Error.StatusCode, envelope.Error.Status, failure.StatusCode)
-		failure.Message = firstNonEmpty(envelope.Error.Message, failure.Message)
-		if envelope.Error.Data != nil {
-			failure.Code = firstNonEmpty(envelope.Error.Data.Code, failure.Code)
-			failure.StatusCode = firstNonZero(envelope.Error.Data.StatusCode, envelope.Error.Data.Status, failure.StatusCode)
-			failure.Message = firstNonEmpty(envelope.Error.Data.Message, failure.Message)
-		}
-	}
-	if envelope.Error == nil && !strings.EqualFold(envelope.Type, "error") {
-		return opencodeStructuredFailure{}, false
-	}
-	return failure, true
-}
-
-func firstNonZero(values ...int) int {
-	for _, value := range values {
-		if value != 0 {
-			return value
-		}
-	}
-	return 0
-}
-
-func classifyOpenCodeStructuredFailure(failure opencodeStructuredFailure) (interfaces.WorkFailureType, bool) {
-	signal := strings.ToLower(strings.Join([]string{failure.Name, failure.Type, failure.Code}, " "))
-	switch {
-	case containsAny(signal, "providerautherror", "authentication_error", "permission_error", "unauthorized", "forbidden"):
-		return interfaces.WorkFailureTypeAuthFailure, true
-	case containsAny(signal, "invalid_request_error", "badrequesterror", "invalidrequesterror"):
-		return interfaces.WorkFailureTypePermanentBadRequest, true
-	case containsAny(signal, "ratelimiterror", "rate_limit_error", "overloaded_error", "quotaexceeded"):
-		return interfaces.WorkFailureTypeThrottled, true
-	case containsAny(signal, "timeouterror", "timeout_error", "etimedout"):
-		return interfaces.WorkFailureTypeTimeout, true
-	case containsAny(signal, "server_error", "internalservererror"):
-		return interfaces.WorkFailureTypeInternalServerError, true
-	}
-	switch {
-	case failure.StatusCode == 401 || failure.StatusCode == 403:
-		return interfaces.WorkFailureTypeAuthFailure, true
-	case failure.StatusCode == 400 || failure.StatusCode == 422:
-		return interfaces.WorkFailureTypePermanentBadRequest, true
-	case failure.StatusCode == 408:
-		return interfaces.WorkFailureTypeTimeout, true
-	case failure.StatusCode == 429:
-		return interfaces.WorkFailureTypeThrottled, true
-	case failure.StatusCode >= 500 && failure.StatusCode <= 599:
-		return interfaces.WorkFailureTypeInternalServerError, true
-	default:
-		return interfaces.WorkFailureTypeUnknown, false
-	}
-}
-
-func lastOpenCodeTextFailure(streams []string, exitCode int) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
-	var found bool
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			trimmed := strings.TrimSpace(line)
-			if !openCodeErrorTextLine(trimmed) {
-				continue
-			}
-			if failure, ok := recognizedOpenCodeTextFailure(trimmed, exitCode); ok {
-				last, found = failure, true
-			}
-		}
-	}
-	if found {
-		return last, true
-	}
-	for _, stream := range streams {
-		trimmed := strings.TrimSpace(stream)
-		if trimmed == "" || strings.Contains(trimmed, "\n") {
-			continue
-		}
-		if failure, ok := recognizedOpenCodeTextFailure(trimmed, exitCode); ok {
-			last, found = failure, true
-		}
-	}
-	return last, found
-}
-
-func openCodeErrorTextLine(line string) bool {
-	normalized := strings.ToLower(line)
-	return strings.HasPrefix(normalized, "error:") || strings.HasPrefix(normalized, "api error:")
-}
-
-func recognizedOpenCodeTextFailure(message string, exitCode int) (ProviderFailureResult, bool) {
-	normalized := strings.ToLower(strings.TrimSpace(message))
-	var reason interfaces.WorkFailureType
-	switch {
-	case exitCode == 124 || containsAny(normalized, "deadline exceeded", "request timed out", "timed out", "timeout"):
-		reason = interfaces.WorkFailureTypeTimeout
-	case containsAny(normalized, "authentication", "login required", "not authenticated", "unauthorized", "forbidden", "api key"):
-		reason = interfaces.WorkFailureTypeAuthFailure
-	case containsAny(normalized, "invalid request", "bad request", "invalid argument", "model not found"):
-		reason = interfaces.WorkFailureTypePermanentBadRequest
-	case containsAny(normalized, "rate limit", "too many requests", "usage limit", "at capacity", "status 429"):
-		reason = interfaces.WorkFailureTypeThrottled
-	case containsAny(normalized, "internal server error", "server error", "status 500", "status 502", "status 503", "status 504"):
-		reason = interfaces.WorkFailureTypeInternalServerError
-	default:
-		return ProviderFailureResult{}, false
-	}
-	return ProviderFailureResult{Reason: reason, Message: openCodeFailureMessage(reason, message)}, true
-}
-
-func openCodeFailureMessage(reason interfaces.WorkFailureType, detail string) string {
-	if reason == interfaces.WorkFailureTypeAuthFailure || reason == interfaces.WorkFailureTypePermanentBadRequest {
-		if sanitized, ok := safeOpenCodeFailureDetail(detail); ok {
-			return sanitized
-		}
-	}
-	switch reason {
-	case interfaces.WorkFailureTypeAuthFailure:
-		return opencodeAuthFailureMessage
-	case interfaces.WorkFailureTypePermanentBadRequest:
-		return opencodeBadRequestFailureMessage
-	case interfaces.WorkFailureTypeThrottled:
-		return opencodeThrottleFailureMessage
-	case interfaces.WorkFailureTypeTimeout:
-		return opencodeTimeoutFailureMessage
-	case interfaces.WorkFailureTypeInternalServerError:
-		return opencodeServerFailureMessage
-	default:
-		return ""
-	}
-}
-
-func safeOpenCodeFailureDetail(detail string) (string, bool) {
-	detail = strings.ToValidUTF8(strings.Join(strings.Fields(detail), " "), "")
-	normalized := strings.ToLower(detail)
-	if detail == "" || containsAny(normalized,
-		"authorization:", "bearer ", "api_key=", "api-key=", `"token":`, "secret=", "sk-", "prompt:", "transcript:",
-	) {
-		return "", false
-	}
-	if len(detail) <= opencodeFailureMessageBytes {
-		return detail, true
-	}
-	end := opencodeFailureMessageBytes
-	for end > 0 && detail[end]&0xc0 == 0x80 {
-		end--
-	}
-	return detail[:end], true
-}
-
-func safeGeminiTextCandidate(line string) string {
-	message := sanitizeGeminiMessage(line)
-	if message == "" || strings.HasPrefix(message, "{") {
-		return ""
-	}
-	normalized := strings.ToLower(message)
-	if isRejectedGeminiMessage(normalized) || !isGeminiErrorSignal(normalized) {
-		return ""
-	}
-	return message
-}
-
-func isGeminiErrorSignal(normalized string) bool {
-	if strings.HasPrefix(normalized, "error:") ||
-		strings.HasPrefix(normalized, "gemini error:") ||
-		strings.HasPrefix(normalized, "fatal") ||
-		strings.HasPrefix(normalized, "failed") ||
-		strings.HasPrefix(normalized, "failure:") ||
-		strings.HasPrefix(normalized, "cannot ") ||
-		strings.HasPrefix(normalized, "could not ") {
-		return true
-	}
-	return containsAny(normalized,
-		"http 4", "http 5", "status 4", "status 5",
-		"unauthenticated", "permission_denied", "resource_exhausted", "resource exhausted",
-		"deadline_exceeded", "timed out", "timeout", "permission denied",
-		"rate limit exceeded", "quota exceeded", "too many requests",
-		"invalid request", "bad request", "service unavailable", "upstream unavailable")
-}
-
-func geminiFailureResult(reason interfaces.WorkFailureType, upstreamMessage string) ProviderFailureResult {
-	message := geminiFixedFailureMessage(reason)
-	if reason == interfaces.WorkFailureTypeAuthFailure || reason == interfaces.WorkFailureTypePermanentBadRequest {
-		if safe := safeGeminiStructuredMessage(upstreamMessage); safe != "" {
-			message = safe
-		}
-	}
-	return ProviderFailureResult{Reason: reason, Message: message}
-}
-
-func geminiFixedFailureMessage(reason interfaces.WorkFailureType) string {
-	switch reason {
-	case interfaces.WorkFailureTypeAuthFailure:
-		return geminiAuthFailureMessage
-	case interfaces.WorkFailureTypePermanentBadRequest:
-		return geminiBadRequestMessage
-	case interfaces.WorkFailureTypeThrottled:
-		return geminiThrottleFailureMessage
-	case interfaces.WorkFailureTypeTimeout:
-		return geminiTimeoutFailureMessage
-	case interfaces.WorkFailureTypeInternalServerError:
-		return geminiServerFailureMessage
-	default:
-		return ""
-	}
-}
-
-func safeGeminiStructuredMessage(message string) string {
-	message = sanitizeGeminiMessage(message)
-	if message == "" {
-		return ""
-	}
-	normalized := strings.ToLower(message)
-	if isRejectedGeminiMessage(normalized) {
-		return ""
-	}
-	return message
-}
-
-func sanitizeGeminiMessage(message string) string {
-	message = strings.Map(func(r rune) rune {
-		if unicode.IsControl(r) {
-			return ' '
-		}
-		return r
-	}, message)
-	message = strings.Join(strings.Fields(message), " ")
-	runes := []rune(message)
-	if len(runes) > geminiFailureMessageRunes {
-		message = string(runes[:geminiFailureMessageRunes])
-	}
-	return message
-}
-
-func isRejectedGeminiMessage(normalized string) bool {
-	if strings.HasPrefix(normalized, "at ") || strings.HasPrefix(normalized, "goroutine ") {
-		return true
-	}
-	return containsAny(normalized,
-		"authorization:", "basic ", "bearer ", "api_key=", "api-key=", "password=", "token=", "secret=", "sk-",
-		"api key:", "credential=", "aiza", "ya29.", "-----begin private key",
-		"customer prompt", "user prompt", "prompt:", "model response", "transcript:",
-		"[debug]", "debug:", "[progress]", "progress:", "traceback", "stack trace",
-		"error report", "report written", "cleanup", "cleaning up", "/tmp/", "/var/tmp/", ".gemini/tmp/")
-}
-
-func tailForGeminiFailureScan(output []byte) string {
-	if len(output) <= geminiFailureScanBytes {
-		return string(output)
-	}
-	return string(output[len(output)-geminiFailureScanBytes:])
-}
