+++ pkg/workers/provider/provider_errors.go
@@ -2,11 +2,8 @@ package provider
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"strings"
-	"unicode"
 
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
@@ -30,861 +27,6 @@ type ProviderError struct {
 type ProviderFailureResult struct {
 	Reason  interfaces.WorkFailureType
 	Message string
-}
-
-const codexFailureMessageBytes = 1024
-
-const opencodeFailureMessageBytes = 512
-
-const (
-	kiroFailureMessageBytes = codexFailureMessageBytes
-	kiroErrorLineScanBytes  = codexErrorLineScanBytes
-)
-
-const (
-	geminiFailureScanBytes    = 64 * 1024
-	geminiFailureMessageRunes = 1024
-)
-
-const (
-	codexAuthFailureMessage       = "Codex authentication failed."
-	codexBadRequestFailureMessage = "Codex rejected the request as invalid."
-	codexGPT56SolUpgradeMessage   = "The 'gpt-5.6-sol' model requires a newer version of Codex. Please upgrade to the latest app or CLI and try again."
-	codexServerFailureMessage     = "Codex encountered a temporary server error."
-	codexThrottleFailureMessage   = "Codex is temporarily unavailable due to usage or capacity limits."
-	codexTimeoutFailureMessage    = "Codex request timed out."
-)
-
-const (
-	geminiAuthFailureMessage         = "Gemini authentication failed."
-	geminiBadRequestMessage          = "Gemini rejected the request."
-	geminiThrottleFailureMessage     = "The provider is rate limited; retry after capacity becomes available."
-	geminiTimeoutFailureMessage      = "Gemini request timed out."
-	geminiServerFailureMessage       = "Gemini encountered a temporary server error."
-	opencodeAuthFailureMessage       = "OpenCode authentication failed."
-	opencodeBadRequestFailureMessage = "OpenCode rejected the request as invalid."
-	opencodeServerFailureMessage     = "OpenCode encountered a temporary server error."
-	opencodeThrottleFailureMessage   = "OpenCode is temporarily unavailable due to usage or capacity limits."
-	opencodeTimeoutFailureMessage    = "OpenCode request timed out."
-)
-
-type codexStructuredFailure struct {
-	Type    string
-	Status  int
-	Message string
-}
-
-type geminiStructuredFailure struct {
-	Type    string
-	Status  string
-	Code    string
-	Message string
-}
-
-// ParseGeminiProviderFailure converts Gemini-owned structured and text failure
-// shapes into one canonical reason/message pair. The final valid structured
-// record wins (stderr is the deterministic cross-stream tie-breaker), followed
-// by recognized stderr text, recognized stdout text, then safe unknown excerpts
-// in that same stderr/stdout order. Both inspected streams and published text
-// are bounded by the Gemini-specific limits above.
-func ParseGeminiProviderFailure(result CommandResult) ProviderFailureResult {
-	streams := []string{
-		tailForGeminiFailureScan(result.Stdout),
-		tailForGeminiFailureScan(result.Stderr),
-	}
-	if failure, ok := lastGeminiStructuredFailure(streams); ok {
-		if failure.Message == "" {
-			failure.Message = fmt.Sprintf("gemini exited with code %d", result.ExitCode)
-		}
-		return failure
-	}
-	if result.ExitCode == 124 {
-		return geminiFailureResult(interfaces.WorkFailureTypeTimeout, "")
-	}
-	stderr := tailForGeminiFailureScan(result.Stderr)
-	if failure, ok := lastGeminiTextFailure(stderr); ok {
-		return failure
-	}
-	stdout := tailForGeminiFailureScan(result.Stdout)
-	if failure, ok := lastGeminiTextFailure(stdout); ok {
-		return failure
-	}
-	if message := lastGeminiUnknownMessage(stderr); message != "" {
-		return ProviderFailureResult{Reason: interfaces.WorkFailureTypeUnknown, Message: message}
-	}
-	if message := lastGeminiUnknownMessage(stdout); message != "" {
-		return ProviderFailureResult{Reason: interfaces.WorkFailureTypeUnknown, Message: message}
-	}
-	return ProviderFailureResult{
-		Reason:  interfaces.WorkFailureTypeUnknown,
-		Message: fmt.Sprintf("gemini exited with code %d", result.ExitCode),
-	}
-}
-
-func lastGeminiStructuredFailure(streams []string) (ProviderFailureResult, bool) {
-	var last geminiStructuredFailure
-	var found bool
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			failure, ok := decodeGeminiStructuredFailure(strings.TrimSpace(line))
-			if !ok {
-				continue
-			}
-			last = failure
-			found = true
-		}
-	}
-	if !found {
-		return ProviderFailureResult{}, false
-	}
-	reason := classifyGeminiFailureSignal(last.Type, last.Status, last.Code, last.Message)
-	if reason != interfaces.WorkFailureTypeUnknown {
-		return geminiFailureResult(reason, last.Message), true
-	}
-	return ProviderFailureResult{
-		Reason:  interfaces.WorkFailureTypeUnknown,
-		Message: safeGeminiStructuredMessage(last.Message),
-	}, true
-}
-
-func decodeGeminiStructuredFailure(payload string) (geminiStructuredFailure, bool) {
-	if !strings.HasPrefix(payload, "{") {
-		return geminiStructuredFailure{}, false
-	}
-	var envelope struct {
-		Type    string          `json:"type"`
-		Status  json.RawMessage `json:"status"`
-		Code    json.RawMessage `json:"code"`
-		Message string          `json:"message"`
-		Error   *struct {
-			Type    string          `json:"type"`
-			Status  json.RawMessage `json:"status"`
-			Code    json.RawMessage `json:"code"`
-			Message string          `json:"message"`
-		} `json:"error"`
-	}
-	if err := json.Unmarshal([]byte(payload), &envelope); err != nil {
-		return geminiStructuredFailure{}, false
-	}
-	if envelope.Error == nil && !isGeminiErrorRecordType(envelope.Type) {
-		return geminiStructuredFailure{}, false
-	}
-	failure := geminiStructuredFailure{
-		Type:    envelope.Type,
-		Status:  geminiJSONScalar(envelope.Status),
-		Code:    geminiJSONScalar(envelope.Code),
-		Message: envelope.Message,
-	}
-	if envelope.Error != nil {
-		if envelope.Error.Type != "" {
-			failure.Type = envelope.Error.Type
-		}
-		if status := geminiJSONScalar(envelope.Error.Status); status != "" {
-			failure.Status = status
-		}
-		if code := geminiJSONScalar(envelope.Error.Code); code != "" {
-			failure.Code = code
-		}
-		if envelope.Error.Message != "" {
-			failure.Message = envelope.Error.Message
-		}
-	}
-	if failure.Type == "" && failure.Status == "" && failure.Code == "" && failure.Message == "" {
-		return geminiStructuredFailure{}, false
-	}
-	return failure, true
-}
-
-func isGeminiErrorRecordType(recordType string) bool {
-	switch strings.ToLower(strings.TrimSpace(recordType)) {
-	case "error", "fatalauthenticationerror", "authenticationerror", "badrequesterror":
-		return true
-	default:
-		return false
-	}
-}
-
-func geminiJSONScalar(raw json.RawMessage) string {
-	if len(raw) == 0 || string(raw) == "null" {
-		return ""
-	}
-	var text string
-	if json.Unmarshal(raw, &text) == nil {
-		return text
-	}
-	return strings.TrimSpace(string(raw))
-}
-
-func lastGeminiTextFailure(stream string) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
-	var found bool
-	for _, line := range strings.Split(stream, "\n") {
-		message := safeGeminiTextCandidate(line)
-		if message == "" {
-			continue
-		}
-		reason := classifyGeminiFailureSignal("", "", "", message)
-		if reason == interfaces.WorkFailureTypeUnknown {
-			continue
-		}
-		last = geminiFailureResult(reason, "")
-		found = true
-	}
-	return last, found
-}
-
-func lastGeminiUnknownMessage(stream string) string {
-	var last string
-	for _, line := range strings.Split(stream, "\n") {
-		if candidate := safeGeminiTextCandidate(line); candidate != "" {
-			last = candidate
-		}
-	}
-	return last
-}
-
-func classifyGeminiFailureSignal(errorType, status, code, message string) interfaces.WorkFailureType {
-	structuredSignals := []string{
-		strings.ToLower(strings.TrimSpace(errorType)),
-		strings.ToLower(strings.TrimSpace(status)),
-		strings.ToLower(strings.TrimSpace(code)),
-	}
-	for _, signal := range structuredSignals {
-		switch signal {
-		case "fatalauthenticationerror", "authenticationerror", "unauthenticated", "permission_denied", "401", "403":
-			return interfaces.WorkFailureTypeAuthFailure
-		case "badrequesterror", "invalid_argument", "400":
-			return interfaces.WorkFailureTypePermanentBadRequest
-		case "resource_exhausted", "ratelimitexceeded", "429":
-			return interfaces.WorkFailureTypeThrottled
-		case "deadline_exceeded", "124", "408":
-			return interfaces.WorkFailureTypeTimeout
-		case "internal", "unavailable", "500", "502", "503", "504":
-			return interfaces.WorkFailureTypeInternalServerError
-		}
-	}
-
-	normalized := strings.ToLower(strings.TrimSpace(message))
-	switch {
-	case containsAny(normalized,
-		"fatalauthenticationerror", "unauthenticated", "permission_denied",
-		"permission denied", "http 401", "status 401", "code 401",
-		"http 403", "status 403", "code 403", "authentication failed",
-		"unauthorized", "forbidden"):
-		return interfaces.WorkFailureTypeAuthFailure
-	case containsAny(normalized,
-		"badrequesterror", "invalid_argument", "invalid argument", "invalid request",
-		"bad request", "http 400", "status 400", "code 400"):
-		return interfaces.WorkFailureTypePermanentBadRequest
-	case containsAny(normalized,
-		"resource_exhausted", "resource exhausted", "ratelimitexceeded", "rate limit",
-		"quota exceeded", "too many requests", "http 429", "status 429", "code 429"):
-		return interfaces.WorkFailureTypeThrottled
-	case containsAny(normalized,
-		"deadline_exceeded", "deadline exceeded", "request timed out", "request timeout",
-		"command timed out", "provider timed out"):
-		return interfaces.WorkFailureTypeTimeout
-	case containsAny(normalized,
-		"internal server error", "service unavailable", "upstream unavailable",
-		"http 500", "status 500", "code 500", "http 502", "status 502", "code 502",
-		"http 503", "status 503", "code 503", "http 504", "status 504", "code 504"):
-		return interfaces.WorkFailureTypeInternalServerError
-	default:
-		return interfaces.WorkFailureTypeUnknown
-	}
-}
-
-type opencodeStructuredFailure struct {
-	Name       string
-	Type       string
-	Code       string
-	StatusCode int
-	Message    string
-}
-
-// ParseOpenCodeProviderFailure deterministically parses bounded OpenCode
-// subprocess output into the canonical provider-failure contract. Recognized
-// structured records take precedence over recognized text diagnostics.
-func ParseOpenCodeProviderFailure(result CommandResult) ProviderFailureResult {
-	streams := []string{
-		tailForCodexErrorScan(result.Stderr),
-		tailForCodexErrorScan(result.Stdout),
-	}
-	if failure, ok := lastOpenCodeStructuredFailure(streams); ok {
-		return failure
-	}
-	if failure, ok := lastOpenCodeTextFailure(streams, result.ExitCode); ok {
-		return failure
-	}
-	if excerpt, ok := lastSafeOpenCodeUnknownExcerpt(streams); ok {
-		return ProviderFailureResult{
-			Reason:  interfaces.WorkFailureTypeUnknown,
-			Message: excerpt,
-		}
-	}
-	return ProviderFailureResult{
-		Reason:  interfaces.WorkFailureTypeUnknown,
-		Message: fmt.Sprintf("opencode exited with code %d", result.ExitCode),
-	}
-}
-
-const (
-	kiroAuthFailureMessage       = "Kiro authentication failed. Sign in again and retry."
-	kiroBadRequestFailureMessage = "Kiro rejected the request as invalid."
-	kiroThrottleFailureMessage   = "Kiro is temporarily unavailable due to usage or capacity limits."
-	kiroTimeoutFailureMessage    = "Kiro request timed out."
-	kiroServerFailureMessage     = "Kiro encountered a temporary service error."
-)
-
-// ParseKiroProviderFailure is the pure Kiro-owned normalization boundary for
-// non-zero CLI exits. It inspects bounded stderr/stdout tails, gives recognized
-// structured records precedence over text, and returns only canonical reasons
-// with product-owned messages for known failures.
-func ParseKiroProviderFailure(result CommandResult) ProviderFailureResult {
-	if result.ExitCode == 124 {
-		return knownKiroFailure(interfaces.WorkFailureTypeTimeout)
-	}
-	streams := []string{
-		tailForKiroErrorScan(result.Stderr),
-		tailForKiroErrorScan(result.Stdout),
-	}
-	if failure, ok := firstKiroStructuredFailure(streams); ok {
-		return failure
-	}
-	if failure, ok := firstKiroTextFailure(streams, result.ExitCode); ok {
-		return failure
-	}
-	if message, ok := firstKiroUnknownFailureExcerpt(streams); ok {
-		return ProviderFailureResult{
-			Reason:  interfaces.WorkFailureTypeUnknown,
-			Message: message,
-		}
-	}
-	return ProviderFailureResult{
-		Reason:  interfaces.WorkFailureTypeUnknown,
-		Message: kiroExitFailureMessage(result.ExitCode),
-	}
-}
-
-func firstKiroStructuredFailure(streams []string) (ProviderFailureResult, bool) {
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			payload := kiroStructuredPayload(line)
-			if payload == "" {
-				continue
-			}
-			reason, ok := decodeKiroStructuredFailure(payload)
-			if ok {
-				return knownKiroFailure(reason), true
-			}
-		}
-	}
-	return ProviderFailureResult{}, false
-}
-
-func kiroStructuredPayload(line string) string {
-	trimmed := strings.TrimSpace(line)
-	for _, prefix := range []string{"ERROR:", "Error:", "KIRO_ERROR:"} {
-		trimmed = strings.TrimSpace(strings.TrimPrefix(trimmed, prefix))
-	}
-	if !strings.HasPrefix(trimmed, "{") {
-		return ""
-	}
-	return trimmed
-}
-
-func decodeKiroStructuredFailure(payload string) (interfaces.WorkFailureType, bool) {
-	var envelope map[string]any
-	if err := json.Unmarshal([]byte(payload), &envelope); err != nil {
-		return interfaces.WorkFailureTypeUnknown, false
-	}
-
-	signals := kiroStructuredSignals(envelope)
-	if nested, ok := envelope["error"].(map[string]any); ok {
-		signals = append(kiroStructuredSignals(nested), signals...)
-	}
-	for _, signal := range signals {
-		if reason := classifyKiroSignal(signal); reason != interfaces.WorkFailureTypeUnknown {
-			return reason, true
-		}
-	}
-	for _, status := range kiroStructuredStatuses(envelope) {
-		if reason := classifyKiroStatus(status); reason != interfaces.WorkFailureTypeUnknown {
-			return reason, true
-		}
-	}
-	if nested, ok := envelope["error"].(map[string]any); ok {
-		for _, status := range kiroStructuredStatuses(nested) {
-			if reason := classifyKiroStatus(status); reason != interfaces.WorkFailureTypeUnknown {
-				return reason, true
-			}
-		}
-	}
-	return interfaces.WorkFailureTypeUnknown, false
-}
-
-func kiroStructuredSignals(record map[string]any) []string {
-	keys := []string{"type", "code", "error_type", "errorType", "name"}
-	signals := make([]string, 0, len(keys))
-	for _, key := range keys {
-		if value, ok := record[key].(string); ok {
-			signals = append(signals, value)
-		}
-	}
-	return signals
-}
-
-func kiroStructuredStatuses(record map[string]any) []int {
-	keys := []string{"status", "status_code", "statusCode"}
-	statuses := make([]int, 0, len(keys))
-	for _, key := range keys {
-		switch value := record[key].(type) {
-		case float64:
-			statuses = append(statuses, int(value))
-		case string:
-			var status int
-			if _, err := fmt.Sscanf(value, "%d", &status); err == nil {
-				statuses = append(statuses, status)
-			}
-		}
-	}
-	return statuses
-}
-
-func classifyKiroSignal(signal string) interfaces.WorkFailureType {
-	normalized := strings.ToLower(strings.TrimSpace(signal))
-	switch {
-	case containsAny(normalized, "authentication", "authorization", "unauthorized", "forbidden", "access_denied", "accessdenied"):
-		return interfaces.WorkFailureTypeAuthFailure
-	case containsAny(normalized, "invalid_request", "invalidrequest", "validation", "bad_request", "badrequest", "invalid_argument"):
-		return interfaces.WorkFailureTypePermanentBadRequest
-	case containsAny(normalized, "rate_limit", "ratelimit", "throttl", "too_many_requests", "capacity", "overloaded"):
-		return interfaces.WorkFailureTypeThrottled
-	case containsAny(normalized, "timeout", "timed_out", "deadline_exceeded"):
-		return interfaces.WorkFailureTypeTimeout
-	case containsAny(normalized, "internal_server", "internalserver", "server_error", "service_unavailable", "serviceunavailable", "api_error"):
-		return interfaces.WorkFailureTypeInternalServerError
-	default:
-		return interfaces.WorkFailureTypeUnknown
-	}
-}
-
-func classifyKiroStatus(status int) interfaces.WorkFailureType {
-	switch {
-	case status == 401 || status == 403:
-		return interfaces.WorkFailureTypeAuthFailure
-	case status == 400 || status == 422:
-		return interfaces.WorkFailureTypePermanentBadRequest
-	case status == 429:
-		return interfaces.WorkFailureTypeThrottled
-	case status == 408 || status == 504:
-		return interfaces.WorkFailureTypeTimeout
-	case status >= 500 && status <= 599:
-		return interfaces.WorkFailureTypeInternalServerError
-	default:
-		return interfaces.WorkFailureTypeUnknown
-	}
-}
-
-func firstKiroTextFailure(streams []string, exitCode int) (ProviderFailureResult, bool) {
-	for _, stream := range streams {
-		lines := strings.Split(stream, "\n")
-		for _, line := range lines {
-			message, ok := kiroTextErrorCandidate(line, len(lines) == 1)
-			if !ok {
-				continue
-			}
-			if reason := classifyKiroTextFailure(message, exitCode); reason != interfaces.WorkFailureTypeUnknown {
-				return knownKiroFailure(reason), true
-			}
-		}
-	}
-	return ProviderFailureResult{}, false
-}
-
-func kiroTextErrorCandidate(line string, singleLine bool) (string, bool) {
-	trimmed := normalizeKiroText(line)
-	if trimmed == "" {
-		return "", false
-	}
-	lower := strings.ToLower(trimmed)
-	if singleLine || strings.HasPrefix(lower, "error:") || strings.HasPrefix(lower, "kiro error:") || strings.HasPrefix(lower, "api error:") {
-		return trimmed, true
-	}
-	return "", false
-}
-
-func normalizeKiroText(value string) string {
-	value = strings.ToValidUTF8(value, " ")
-	value = strings.Map(func(r rune) rune {
-		if unicode.IsControl(r) {
-			return ' '
-		}
-		return r
-	}, value)
-	return strings.Join(strings.Fields(value), " ")
-}
-
-// firstKiroUnknownFailureExcerpt considers stderr before stdout and selects the
-// first explicit error record in that stream. This precedence is intentional:
-// Kiro writes invocation failures to stderr, while stdout can contain model
-// output or an echoed prompt. Conservative rejection keeps ambiguous records
-// on the fixed exit-code fallback instead of risking customer-data exposure.
-func firstKiroUnknownFailureExcerpt(streams []string) (string, bool) {
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			detail, ok := kiroUnknownErrorDetail(line)
-			if !ok {
-				continue
-			}
-			message := truncateKiroFailureMessage("Kiro error: " + detail)
-			if message != "Kiro error:" {
-				return message, true
-			}
-		}
-	}
-	return "", false
-}
-
-func kiroUnknownErrorDetail(line string) (string, bool) {
-	normalized := normalizeKiroText(line)
-	lower := strings.ToLower(normalized)
-	prefixes := []string{"kiro_error:", "kiro error:", "api error:", "error:"}
-	for _, prefix := range prefixes {
-		if !strings.HasPrefix(lower, prefix) {
-			continue
-		}
-		detail := strings.TrimSpace(normalized[len(prefix):])
-		if detail == "" || unsafeKiroUnknownDetail(detail) {
-			return "", false
-		}
-		return detail, true
-	}
-	return "", false
-}
-
-func unsafeKiroUnknownDetail(detail string) bool {
-	lower := strings.ToLower(detail)
-	if strings.HasPrefix(detail, "{") || strings.HasPrefix(detail, "[") || strings.Contains(detail, "=") {
-		return true
-	}
-	return containsAny(lower,
-		"prompt", "transcript", "response draft", "model output", "user message",
-		"request body", "request payload", "input content", "output content",
-		"progress", "cleanup", "credential", "secret", "password",
-		"api key", "api_key", "access key", "authorization", "bearer ",
-		"access token", "refresh token", "session token", "environment", "env var",
-	)
-}
-
-func truncateKiroFailureMessage(message string) string {
-	if len(message) <= kiroFailureMessageBytes {
-		return message
-	}
-	end := 0
-	for index := range message {
-		if index > kiroFailureMessageBytes {
-			break
-		}
-		end = index
-	}
-	return strings.TrimSpace(message[:end])
-}
-
-func classifyKiroTextFailure(message string, exitCode int) interfaces.WorkFailureType {
-	normalized := strings.ToLower(message)
-	switch {
-	case exitCode == 124, containsAny(normalized, "request timed out", "operation timed out", "timeout waiting", "deadline exceeded"):
-		return interfaces.WorkFailureTypeTimeout
-	case containsAny(normalized, "authentication required", "authentication failed", "authorization failed", "not authorized", "unauthorized", "forbidden", "sign in required", "login required"):
-		return interfaces.WorkFailureTypeAuthFailure
-	case containsAny(normalized, "invalid request", "invalid input", "invalid argument", "bad request", "validation failed"):
-		return interfaces.WorkFailureTypePermanentBadRequest
-	case containsAny(normalized, "rate limit", "too many requests", "throttl", "capacity limit", "at capacity", "resource exhausted"):
-		return interfaces.WorkFailureTypeThrottled
-	case containsAny(normalized, "internal server error", "temporary service error", "service unavailable", "unexpected status 500", "unexpected status 502", "unexpected status 503"):
-		return interfaces.WorkFailureTypeInternalServerError
-	default:
-		return interfaces.WorkFailureTypeUnknown
-	}
-}
-
-func knownKiroFailure(reason interfaces.WorkFailureType) ProviderFailureResult {
-	message := ""
-	switch reason {
-	case interfaces.WorkFailureTypeAuthFailure:
-		message = kiroAuthFailureMessage
-	case interfaces.WorkFailureTypePermanentBadRequest:
-		message = kiroBadRequestFailureMessage
-	case interfaces.WorkFailureTypeThrottled:
-		message = kiroThrottleFailureMessage
-	case interfaces.WorkFailureTypeTimeout:
-		message = kiroTimeoutFailureMessage
-	case interfaces.WorkFailureTypeInternalServerError:
-		message = kiroServerFailureMessage
-	}
-	if len(message) > kiroFailureMessageBytes {
-		message = message[:kiroFailureMessageBytes]
-	}
-	return ProviderFailureResult{Reason: reason, Message: message}
-}
-
-func tailForKiroErrorScan(output []byte) string {
-	if len(output) <= kiroErrorLineScanBytes {
-		return string(output)
-	}
-	return string(output[len(output)-kiroErrorLineScanBytes:])
-}
-
-func kiroExitFailureMessage(exitCode int) string {
-	return fmt.Sprintf("kiro-cli exited with code %d", exitCode)
-}
-
-// ParseCodexProviderFailure deterministically parses bounded subprocess output
-// into the canonical provider-failure contract. Each stream is limited by
-// codexErrorLineScanBytes before parsing, and returned messages are limited by
-// codexFailureMessageBytes.
-func ParseCodexProviderFailure(result CommandResult) ProviderFailureResult {
-	streams := []string{
-		tailForCodexErrorScan(result.Stderr),
-		tailForCodexErrorScan(result.Stdout),
-	}
-	if failure, ok := lastCodexStructuredFailure(streams); ok {
-		return failure
-	}
-
-	if failure, ok := lastCodexTextFailure(streams, result.ExitCode); ok {
-		return ProviderFailureResult{
-			Reason:  failure.Reason,
-			Message: failure.Message,
-		}
-	}
-
-	return ProviderFailureResult{
-		Reason:  classifyCodexExitFailure(result.ExitCode),
-		Message: codexExitFailureMessage(result.ExitCode),
-	}
-}
-
-func lastCodexStructuredFailure(streams []string) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
-	var found bool
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			payload, ok := codexErrorPayload(line)
-			if !ok || !strings.HasPrefix(payload, "{") {
-				continue
-			}
-			failure, ok := decodeCodexStructuredFailure(payload)
-			if !ok {
-				continue
-			}
-			reason, recognized := classifyCodexStructuredSignal(failure.Type, failure.Status)
-			if recognized {
-				last = ProviderFailureResult{
-					Reason:  reason,
-					Message: codexStructuredFailureMessage(failure.Message, reason),
-				}
-				found = true
-			}
-		}
-	}
-	return last, found
-}
-
-// codexStructuredFailureMessage publishes only positively audited text. Other
-// structured messages can contain prompts, transcripts, cleanup paths, or
-// credentials, so recognized reasons use fixed customer-visible messages.
-func codexStructuredFailureMessage(message string, reason interfaces.WorkFailureType) string {
-	message = strings.Join(strings.Fields(message), " ")
-	if reason == interfaces.WorkFailureTypePermanentBadRequest && message == codexGPT56SolUpgradeMessage {
-		return message
-	}
-	return codexTextFailureMessage(reason)
-}
-
-func codexErrorPayload(line string) (string, bool) {
-	trimmed := strings.TrimSpace(line)
-	if !strings.HasPrefix(trimmed, "ERROR:") {
-		return "", false
-	}
-	return strings.TrimSpace(strings.TrimPrefix(trimmed, "ERROR:")), true
-}
-
-func decodeCodexStructuredFailure(payload string) (codexStructuredFailure, bool) {
-	var envelope struct {
-		Type    string `json:"type"`
-		Status  int    `json:"status"`
-		Message string `json:"message"`
-		Error   *struct {
-			Type    string `json:"type"`
-			Message string `json:"message"`
-		} `json:"error"`
-	}
-	if err := json.Unmarshal([]byte(payload), &envelope); err != nil {
-		return codexStructuredFailure{}, false
-	}
-	if envelope.Error == nil && envelope.Type == "" && envelope.Status == 0 {
-		return codexStructuredFailure{}, false
-	}
-
-	failure := codexStructuredFailure{
-		Type:    envelope.Type,
-		Status:  envelope.Status,
-		Message: envelope.Message,
-	}
-	if envelope.Error != nil {
-		if envelope.Error.Type != "" {
-			failure.Type = envelope.Error.Type
-		}
-		if envelope.Error.Message != "" {
-			failure.Message = envelope.Error.Message
-		}
-	}
-	return failure, true
-}
-
-func lastCodexTextFailure(streams []string, exitCode int) (ProviderFailureResult, bool) {
-	var last ProviderFailureResult
-	var found bool
-	for _, stream := range streams {
-		for _, line := range strings.Split(stream, "\n") {
-			payload, isError := codexErrorPayload(line)
-			if !isError || strings.HasPrefix(payload, "{") {
-				continue
-			}
-			if failure, ok := recognizedCodexTextFailure(payload, exitCode); ok {
-				last, found = failure, true
-			}
-		}
-	}
-	if found {
-		return last, true
-	}
-
-	// A single unprefixed line is an existing Codex diagnostic shape (for
-	// example transport timeouts). Multi-line output is treated as transcript
-	// or header noise unless it contains an explicit ERROR record.
-	for _, stream := range streams {
-		trimmed := strings.TrimSpace(stream)
-		if trimmed == "" || strings.Contains(trimmed, "\n") {
-			continue
-		}
-		if failure, ok := recognizedCodexTextFailure(trimmed, exitCode); ok {
-			last, found = failure, true
-		}
-	}
-	return last, found
-}
-
-// recognizedCodexTextFailure accepts only audited Codex diagnostic shapes and
-// returns fixed customer-visible text. Unknown ERROR lines may contain echoed
-// prompts, transcripts, cleanup paths, or credentials, so they are never used
-// as message excerpts.
-func recognizedCodexTextFailure(message string, exitCode int) (ProviderFailureResult, bool) {
-	normalized := strings.ToLower(strings.TrimSpace(message))
-	reason := classifyRecognizedCodexTextFailure(normalized, exitCode)
-	if reason == interfaces.WorkFailureTypeUnknown {
-		return ProviderFailureResult{}, false
-	}
-	return ProviderFailureResult{
-		Reason:  reason,
-		Message: codexTextFailureMessage(reason),
-	}, true
-}
-
-func classifyRecognizedCodexTextFailure(message string, exitCode int) interfaces.WorkFailureType {
-	switch {
-	case exitCode == 124,
-		strings.HasPrefix(message, "context deadline exceeded"),
-		strings.HasPrefix(message, "command timed out"),
-		strings.HasPrefix(message, "request timed out"),
-		strings.HasPrefix(message, "provider timeout"),
-		strings.HasPrefix(message, "context canceled after command timed out"):
-		return interfaces.WorkFailureTypeTimeout
-	case strings.HasPrefix(message, "unexpected status 401"),
-		strings.HasPrefix(message, "unexpected status 403"):
-		return interfaces.WorkFailureTypeAuthFailure
-	case strings.HasPrefix(message, "unexpected status 400"):
-		return interfaces.WorkFailureTypePermanentBadRequest
-	case strings.HasPrefix(message, "unexpected status 429"),
-		strings.HasPrefix(message, "you've hit your usage limit"),
-		strings.HasPrefix(message, "selected model is at capacity"):
-		return interfaces.WorkFailureTypeThrottled
-	case strings.HasPrefix(message, "unexpected status 500"),
-		strings.HasPrefix(message, "unexpected status 502"),
-		strings.HasPrefix(message, "unexpected status 503"),
-		strings.HasPrefix(message, "unexpected status 504"),
-		message == codexHighDemandTemporaryErrorsNeedle:
-		return interfaces.WorkFailureTypeInternalServerError
-	default:
-		return interfaces.WorkFailureTypeUnknown
-	}
-}
-
-func codexTextFailureMessage(reason interfaces.WorkFailureType) string {
-	switch reason {
-	case interfaces.WorkFailureTypeAuthFailure:
-		return codexAuthFailureMessage
-	case interfaces.WorkFailureTypePermanentBadRequest:
-		return codexBadRequestFailureMessage
-	case interfaces.WorkFailureTypeThrottled:
-		return codexThrottleFailureMessage
-	case interfaces.WorkFailureTypeInternalServerError:
-		return codexServerFailureMessage
-	case interfaces.WorkFailureTypeTimeout:
-		return codexTimeoutFailureMessage
-	default:
-		return ""
-	}
-}
-
-func classifyCodexExitFailure(exitCode int) interfaces.WorkFailureType {
-	if exitCode == 124 {
-		return interfaces.WorkFailureTypeTimeout
-	}
-	if exitCode == codexWindowsProcessFailureExitCode {
-		return interfaces.WorkFailureTypeInternalServerError
-	}
-	return interfaces.WorkFailureTypeUnknown
-}
-
-func classifyCodexStructuredSignal(errorType string, status int) (interfaces.WorkFailureType, bool) {
-	normalizedType := strings.ToLower(strings.TrimSpace(errorType))
-	switch normalizedType {
-	case "authentication_error", "permission_error":
-		return interfaces.WorkFailureTypeAuthFailure, true
-	case "invalid_request_error":
-		return interfaces.WorkFailureTypePermanentBadRequest, true
-	case "rate_limit_error", "overloaded_error":
-		return interfaces.WorkFailureTypeThrottled, true
-	case "api_error", "server_error":
-		return interfaces.WorkFailureTypeInternalServerError, true
-	case "", "error":
-		// Generic Codex envelopes carry the provider classification in status.
-	default:
-		// An explicit unknown type can identify cleanup or diagnostic records;
-		// its status must not let it override a recognized provider failure.
-		return interfaces.WorkFailureTypeUnknown, false
-	}
-
-	switch {
-	case status == 401 || status == 403:
-		return interfaces.WorkFailureTypeAuthFailure, true
-	case status == 400:
-		return interfaces.WorkFailureTypePermanentBadRequest, true
-	case status == 429:
-		return interfaces.WorkFailureTypeThrottled, true
-	case status >= 500 && status <= 599:
-		return interfaces.WorkFailureTypeInternalServerError, true
-	case status == 408:
-		return interfaces.WorkFailureTypeTimeout, true
-	default:
-		return interfaces.WorkFailureTypeUnknown, false
-	}
-}
-func codexExitFailureMessage(exitCode int) string {
-	return fmt.Sprintf("codex exited with code %d", exitCode)
 }
 
 func NewProviderError(errorType interfaces.WorkFailureType, message string, cause error) *ProviderError {
